Display age instead of birthdate

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('lastname');
             $table->string('github_username')->nullable();
             $table->date('birthdate');
-            $table->tinyInteger('gender')->unsigned();
+            $table->integer('gender')->unsigned();
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
